Entidades y builders de las últimas 6 tablas.

+++ Aplicacion Desktop/GestionCommon/Entidades/Rol.cs
@@ -7,7 +7,7 @@ namespace GestionCommon.Entidades
 {
     public class Rol
     {
-        public int IdRol { get; set; }
+        public decimal IdRol { get; set; }
         public string Nombre { get; set; }
         public bool Habilitado { get; set; }
     }
