+++ Aplicacion Desktop/GestionDomain/IEntidadDomain.cs
@@ -10,9 +10,9 @@ namespace GestionDomain
     {
         IEntidadDAL<T> DAL { get; }
 
-        T Obtener(int id);
+        T Obtener(decimal id);
         IList<T> ObtenerTodos();
-        void Borrar(int id);
+        void Borrar(decimal id);
         T Modificar(T entidad);
         T Crear(T entidad);
     }
